--- src/pages/ProductDetail/DetailData/DetailImgs.js
@@ -1,14 +1,0 @@
-export const DetailImgs = [
-  {
-    img: '/images/chamara.jpg',
-    desc: 'wine1',
-  },
-  {
-    img: '/images/chamara2.jpg',
-    desc: 'wine1',
-  },
-  {
-    img: '/images/wine_label.png',
-    desc: 'wine1',
-  },
-];
