+++ src/pages/ProductDetail/ProductDetail.scss
@@ -5,140 +5,4 @@
   margin: 0 auto;
   padding-left: 100px;
   padding-top: 19px;
-  // background-color: skyblue;
-
-  .slideCurrentImgWrap {
-    width: 62%;
-    height: 615px;
-
-    .breadcrumbs {
-      margin-bottom: 20px;
-      line-height: 1.3;
-      font-size: 14px;
-    }
-
-    .slideCurrent {
-      width: 100%;
-      height: 100%;
-      object-fit: cover;
-    }
-
-    .productsContent {
-      position: absolute;
-      top: 55px;
-      left: 0;
-
-      ul {
-        li {
-          .productImage {
-            width: 80px;
-            height: 80px;
-            margin-bottom: 12px;
-            opacity: 0.6;
-            cursor: pointer;
-            transition: all 0.3s ease-in-out;
-            object-fit: cover;
-          }
-
-          .selected {
-            transition: all 0.3s ease-in-out;
-            opacity: 1;
-          }
-        }
-      }
-    }
-
-    .slideCurrentImageFoot {
-      padding: 20px 15px;
-      border-left: 1px solid #e9e9e9;
-
-      p {
-        font-size: 14px;
-        span {
-          .underLine {
-            text-decoration: underline;
-            cursor: pointer;
-          }
-        }
-      }
-    }
-  }
-
-  .productSingleSide {
-    border-left: 1px solid #e9e9e9;
-    border-right: 1px solid #e9e9e9;
-    padding-top: 30px;
-    width: 38%;
-
-    .productHead {
-      margin-bottom: 20px;
-      padding: 0 10px 17px 10px;
-
-      div {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-bottom: 10px;
-        h2 {
-          font-size: 26px;
-          line-height: 1.3;
-          font-weight: bold;
-        }
-
-        .productPrice {
-          font-size: 20px;
-          text-transform: uppercase;
-        }
-      }
-
-      .afterpayParagraph {
-        text-align: right;
-        font-size: 13px;
-        letter-spacing: 1px;
-        padding: 5px 0;
-      }
-
-      .productType {
-        font-weight: bold;
-        font-size: 16px;
-        line-height: 1.19;
-      }
-    }
-
-    .formRow {
-      display: flex;
-      border: 1px solid #e9e9e9;
-      padding: 0 15px;
-      color: #404040;
-      font-size: 14px;
-
-      .productLabel {
-        width: 150px;
-      }
-
-      .productInfo {
-        display: flex;
-        width: 100%;
-        justify-content: space-between;
-        align-items: center;
-        cursor: pointer;
-
-        p {
-          .sizeInput {
-            border: none;
-            outline: none;
-            cursor: pointer;
-          }
-        }
-
-        i {
-          font-size: 10px;
-        }
-      }
-
-      p {
-        line-height: 52px;
-      }
-    }
-  }
 }
