+++ src/pages/ProductDetail/ProductDetailContent/ProductDetailContent.scss
@@ -1,0 +1,56 @@
+.productDetailContent {
+  width: 62%;
+  height: 615px;
+
+  .breadcrumbs {
+    margin-bottom: 20px;
+    line-height: 1.3;
+    font-size: 14px;
+  }
+
+  .slideCurrent {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+
+  .productsContent {
+    position: absolute;
+    top: 55px;
+    left: 0;
+
+    ul {
+      li {
+        .productImage {
+          width: 80px;
+          height: 80px;
+          margin-bottom: 12px;
+          opacity: 0.6;
+          cursor: pointer;
+          transition: all 0.3s ease-in-out;
+          object-fit: cover;
+        }
+
+        .selected {
+          transition: all 0.3s ease-in-out;
+          opacity: 1;
+        }
+      }
+    }
+  }
+
+  .slideCurrentImageFoot {
+    padding: 20px 15px;
+    border-left: 1px solid #e9e9e9;
+
+    p {
+      font-size: 14px;
+      span {
+        .underLine {
+          text-decoration: underline;
+          cursor: pointer;
+        }
+      }
+    }
+  }
+}
