+++ src/pages/ProductDetail/ProductDetailSide/ProductDetailSide.scss
@@ -1,0 +1,77 @@
+.productDetailSide {
+  border-left: 1px solid #e9e9e9;
+  border-right: 1px solid #e9e9e9;
+  padding-top: 30px;
+  width: 38%;
+
+  .productHead {
+    margin-bottom: 20px;
+    padding: 0 10px 17px 10px;
+
+    div {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      margin-bottom: 10px;
+      h2 {
+        font-size: 26px;
+        line-height: 1.3;
+        font-weight: bold;
+      }
+
+      .productPrice {
+        font-size: 20px;
+        text-transform: uppercase;
+      }
+    }
+
+    .afterpayParagraph {
+      text-align: right;
+      font-size: 13px;
+      letter-spacing: 1px;
+      padding: 5px 0;
+    }
+
+    .productType {
+      font-weight: bold;
+      font-size: 16px;
+      line-height: 1.19;
+    }
+  }
+
+  .formRow {
+    display: flex;
+    border: 1px solid #e9e9e9;
+    padding: 0 15px;
+    color: #404040;
+    font-size: 14px;
+
+    .productLabel {
+      width: 150px;
+    }
+
+    .productInfo {
+      display: flex;
+      width: 100%;
+      justify-content: space-between;
+      align-items: center;
+      cursor: pointer;
+
+      p {
+        .sizeInput {
+          border: none;
+          outline: none;
+          cursor: pointer;
+        }
+      }
+
+      i {
+        font-size: 10px;
+      }
+    }
+
+    p {
+      line-height: 52px;
+    }
+  }
+}
